Fetch and expose events data from Github
First, we need to add Jackson-annotated classes in order to
deserialize JSON API responses from api.github.com.
(See `RepositoryEvent`, `Actor`, `Issue`).

We then create a `RestClientTest` to test our new GithubClient
and its interactions with the GitHub API. For that, we can use
a sample response API `event.json` and test our client against
that.

We can then expose this with a Controller, on the
`"/events/{repoName}"` path.