Use GitHub API token
Since the GitHub API is rate-limited, we should create a
`GithubProperties` properties class bound to the "github"
configuration namespace. It is not only possible to get IDE
support with the Spring Boot annotation processor for confituration
properties, but also properties validation.

Finally, we can use that property in a RestTemplate interceptor
to add an authentication request header to each request.