Configure Security
We'd like to add now an admin page to manage the projects displayed
on the dashboard page. But we don't want anyone to access that.

For that, we can add the Spring Boot Security Starter and restart
our application: everything is now secured using a randomly
generated user account.

We can customize this by adding a Security configuration class that:

* Configures a standard user, and an admin with specific roles
* Restrict our admin page to ADMINs only
* Allow the events API to registered users only
* Allow the dashboard page to all
* Secure the actuator endpoints using the Actuator config infrastructure